Fix several bugs, raise error in validate_goal
Instead of returning a boolean + message in `validate_goal`, rather
raise an error if the goal is invalid.  This prevents accidental wrong
use.  When returning a tuple of boolean and string, there is a high risk
that people will do `if validate_goal(...):` which will always be True,
even if the first element of the returned tuple is False.