Move robot to rest position during shutdown
Use the `shutdown()` method of the driver to move the robot back to the
initial position before stopping completely.

This ensures the robot is moved to a good position to start the next run
while it is still initialized (i.e. it can be detected if it is blocked
and does not reach the target position).