Split demo_real_finger.py
Split demo_real_finger.py into separate demos for torque and position
control.  This makes them shorter and less convoluted and thus hopefully
easier to follow.  This serves as the main example of the robot
interface and as entry point for all new users.  Therefore it should be
as simple and easy to understand as possible.