Separate scripts for data and robot backend
Provide two separate scripts `trifinger_data_backend.py` and
`trifinger_robot_backend.py` as an alternative to the combined
`trifinger_backend.py`.

The "data backend" only runs the multi-process robot/sensor data and
loggers and the actual robot/sensor backends are run in the "robot
backend" script.  This way there is better control when shutting down
things as the robot data can still stay alive for the loggers and user
code to finish when the robot backend has already shut down.

Use ROS publishers to notify when the backends are ready and services to
shut them down cleanly.