Add homing_method to Config and update homing()
Add a new configuration parameter "homing_method" to specify the
homing method.  This makes "homing_with_index" obsolete and the latter
will cause an error from now on (so users see that they should update
their config).

Possible homing methods are:

* NONE: Do not do any homing.
* CURRENT_POSITION: Home at current position (previously the behaviour
  of `has_endstop = false, homing_with_index = false`).
* NEXT_INDEX: Starting from the current position, home at the next
  encoder index (previously `has_endstop = false, homing_with_index =
  true`).
* ENDSTOP: Push against end stop and home there (previously `has_endstop
  = true, homing_with_index = false`).
* ENDSTOP_INDEX: First go to end stop, then search index (previously
  `has_endstop = true, homing_with_index = true`).
* ENDSTOP_RELEASE: Go to index, release motors and home there
* (previously implemented in `home_at_stable_endstop()`).

If 'homing_method' is not set, a backward-compatible default is chosen
based on the value of 'has_endstop'.

The method `home_at_stable_endstop()` is removed again and instead its
functionality is incorporated in the old `homing()` method
(homing_method = ENDSTOP_RELEASE).