Log run duration of the robot
Add a configuration option "run_duration_logfiles" where a list of
filenames can be given.  In the shutdown method of the robot append a
line to each specified file containing the current time and the number
of actions that have been sent to the robot.

The purpose of this log is get an idea of how long the mechanics of the
robot last before things break.  The parameter expects a list of log
files so that separate logs for the separate manipulators of TriFinger
can be maintained.