Add "full range" mode to FingerPro endurance test
Add an alternative mode to the FingerPro endurance test (activated by
setting the `--full-range` flag).  In this mode position limits are
ignored and the finger moves on a hard-coded trajectory that uses the
full mechanical range of the joints.

In this mode hard-coded positions are used instead of random ones to
avoid collisions with the surroundings.