Enable use of config objects in the Robot class
Extend the `Robot` class, so that the constructor can handle both config
file paths and config objects.  This makes it easier to initialise a
robot with a config object.

To be consistent in both cases, `Robot.config` is changed to be an
object that can be accessed using attributes instead of being a
dictionary.  Update fingerpro_endurance_test.py accordingly.