evaluate_log: Avoid reward re-computation for dice task
Reward computation of the "rearrange_dice" task is rather
time-consuming and thus evaluation takes a lot of time if the reward is
computed for each robot step.  Since the reward is only based on the
camera observations, rewards can be cached and reused in later robot
steps with the same camera observation.

On my machine this change speeds up the evaluation of one log from
around 4:15min to 12s.