post_submission: Set edge mean threshold to 10
After changing the rest position, less of the robot is visible to the
cameras while sampling images for the sharpness test.  Due to this,
fewer edges are found, resulting in a lower "edge mean" value, even if
the sharpness of the camera is good.
Thus reduce the threshold to 10 to avoid false positives.