trifinger_backend: Get camera rate from driver
Instead of hard-coding it in the script, use the new `get_sensor_info()`
method of the driver to get the camera frame rate.

The frame rate is specified for each of the three cameras, but since
they are syncronised, they are all the same, so we can simply take the
value from the first camera in the list.